Fix for issue#44
Problem was that puzzle_save_common was writing the file
with print(), rather than write.  Print adds a newline.
Cumulative opens and saves of the same puzzle were
adding newlines each time, which showed up as a difference
in the initial and current json.

I had previous used strip() on the json string, but
this is not necessary now.