Fix for issue #81
Puzzle title was not always being displayed because
the puzzle.html file is not always rendered from
`webapp.puzzle_screen()`.  This means that the
`{{ puzzletitle }}` template variable was not being
set, so the conditional code in the tag that
shows the title wasn't following the True path.

It turns out that `webapp.edit_word_screen()`
was rendering `puzzle.html` itself.

The fix was to have `edit_word_screen()` finish
with
```python
return redirect(url_for('puzzle_screen'))
```