Use sha256 of initial json, not entire string
In order to check for "close without saving?",
the system currently saves the JSON image of the
entire puzzle or grid in the session and compares
that to the current JSON image of the object when
the close function is called.

This is not necessary and consumes a lot of session
space.  It is just as effective to store the
**sha1** of the JSON images for comparison.  This
value is only 32 bytes long.